% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydiff.R
\name{get_diff}
\alias{get_diff}
\title{Get the diff result from a shinydiff widget}
\usage{
get_diff(outputId, session = NULL, as = c("data.frame", "list"))
}
\arguments{
\item{outputId}{The outputId used in shinydiffOutput.}

\item{session}{The shiny session (defaults to current reactive domain).}

\item{as}{(optional) "data.frame" (default) or "list" for return type.}
}
\value{
A data.frame or list of diff segments, or NULL if not available.
}
\description{
Get the diff result from a shinydiff widget
}
