% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydiff.R
\name{shinydiff}
\alias{shinydiff}
\title{shinydiff htmlwidget}
\usage{
shinydiff(
  text1,
  text2,
  diffType = c("words", "chars", "lines", "sentences", "trimmedlines", "css"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{text1}{First text string}

\item{text2}{Second text string}

\item{diffType}{Type of diff: "chars", "words", "lines", "sentences", "trimmedlines", or "css"}

\item{width}{Widget width}

\item{height}{Widget height}

\item{elementId}{Widget element id}
}
\description{
shinydiff htmlwidget
}
