% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydiff.R
\name{shinydiffModuleServer}
\alias{shinydiffModuleServer}
\title{Shiny module server for shinydiff}
\usage{
shinydiffModuleServer(id, text1, text2)
}
\arguments{
\item{id}{Module id}

\item{text1}{Reactive for first text}

\item{text2}{Reactive for second text}
}
\description{
Shiny module server for shinydiff
}
