% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydiff.R
\name{shinydiffOutput}
\alias{shinydiffOutput}
\title{Shiny bindings for shinydiff}
\usage{
shinydiffOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{Must be a valid CSS unit}
}
\description{
Shiny bindings for shinydiff
}
