% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydiff.R
\name{renderShinydiff}
\alias{renderShinydiff}
\title{Render shinydiff widget}
\usage{
renderShinydiff(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a shinydiff widget.}

\item{env}{The environment in which to evaluate expr. Defaults to parent.frame().}

\item{quoted}{Logical, whether expr is a quoted expression (default FALSE).}
}
\value{
A Shiny render function for use in server logic.
}
\description{
Render shinydiff widget
}
