\name{shinydiffModuleServer}
\alias{shinydiffModuleServer}
\title{Shiny module server for shinydiff}
\description{Server function for the shinydiff Shiny module.}
\usage{shinydiffModuleServer(id, text1, text2)}
\arguments{
  \item{id}{Module id}
  \item{text1}{Reactive for first text}
  \item{text2}{Reactive for second text}
}
\examples{
# See README for example usage in a Shiny app
}
